requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
discord==1.0.1
discord.py==1.5.1
idna==2.10
imgurpython==1.1.7
multidict==4.7.6
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
praw==7.1.0
prawcore==1.5.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.1
six==1.15.0
soupsieve==2.1
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0
yarl==1.5.1
